24.CNN/CNN_Project_DogsCatsClassification/requirements.txt
tensorflow==2.15.0
streamlit==1.37.1
numpy
pandas
Pillow
matplotlib
streamlit